Fix smoke test for transformers 5.x API
apply_chat_template now returns a dict (input_ids + attention_mask), so
generate() needs **inputs and the prompt length read from
inputs['input_ids']. Also rename torch_dtype -> dtype in the model loader
to clear the deprecation warning. Model loads and downloads fine on MPS;
this was the only blocker to a green smoke test.

+++ src/mvm/model.py
@@ -30,7 +30,7 @@ def load_model(device: str | None = None):
     model = AutoModelForCausalLM.from_pretrained(
         config.MODEL_ID,
         revision=config.MODEL_REVISION,
-        torch_dtype=dtype,
+        dtype=dtype,  # transformers 5.x renamed torch_dtype -> dtype
         attn_implementation=config.ATTN_IMPLEMENTATION,
     )
     model.to(device)
